Refactor: separate selected price component
